// src/ui/types.ts
/**
 * @type type that defines the possible responses of the entityMenu
 */

export type EntityMenu =
  | "characters"
  | "dimensions"
  | "species"
  | "locations"
  | "inventions";
  
/**
 * @type type that defines the possible responses of the MainOption
 */
export type MainOption =
  | "char"
  | "dim"
  | "spec"
  | "loc"
  | "inv"
  | "events"
  | "search"
  | "reports"
  | "exit";